+++ src/dexterity/localrolesfield/subscriber.py
@@ -85,3 +85,17 @@ def related_change_on_addition(obj, event):
         if name not in fti_config:
             continue
         related_role_addition(obj, get_state(obj), fti_config[name], name)
+
+
+def related_change_on_removal(obj, event):
+    """ Set local roles on related objects after removal """
+    fti_config = fti_configuration(obj)
+    fti = getUtility(IDexterityFTI, name=obj.portal_type)  # Must be returned by fti_configuration
+    for (name, f) in get_localrole_fields(fti):
+        if name not in fti_config:
+            continue
+        # We have to remove the configuration linked to deleted object
+        # There is a problem in Plone 4.3. The event is notified before the confirmation and after too.
+        # The action could be cancelled: we can't know this !! Resolved in Plone 5...
+        # We choose to update related objects anyway !!
+        related_role_removal(obj, get_state(obj), fti_config[name], name)